% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/redmunds.R
\docType{package}
\name{redmunds}
\alias{redmunds}
\alias{redmunds-package}
\title{redmunds: A package for getting car data from the Edmunds.com api}
\description{
redmunds: A package for getting car data from the Edmunds.com api
}

