% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makes.R
\name{get_makes}
\alias{get_makes}
\title{get a list of all car makes}
\usage{
get_makes(key = getOption("edmunds_api_key"))
}
\arguments{
\item{key}{API key. If you have your own you can supply it,
           otherwise the default package key will be used}
}
\value{
A data frame of car makes, with additional columns 'nice_name'
   and 'make_id' to use in other function calls
}
\description{
get a list of all car makes
}

